Hangman Winform Layout
Created hangman layout on winform. 
The groupbox containing the buttons is where the user will be making their guess. The user will be able to submit a letter or a word. 
The groubox used to the left of the form will display where the hangman is. 
The groupbox used above the buttons will contain the underscores along with a label to tell the user how long the word is.